Starting the service container and move files around.
