+++ bifrost/src/ioc/ServiceContainer.py
@@ -1,0 +1,16 @@
+class Container:
+
+    services = {}
+
+    def __init__(self):
+        self.reset_services()
+
+    def reset_services(self):
+        pass
+
+    def get_service(self):
+        pass
+
+    def set_service(self):
+        pass
+
