+++ bifrost_events/tests/test_accommodation_event.py
@@ -1,4 +1,4 @@
-from bifrost.src.TestUtils.common_test_utils import BaseTestUtils
+from bifrost.src.CommonTestUtils import BaseTestUtils
 from bifrost_events.models import Accommodation
 from datetime import datetime
