+++ bifrost_location/tests/test_location.py
@@ -1,5 +1,5 @@
 from django.db.utils import IntegrityError
-from bifrost.src.TestUtils.common_test_utils import BaseTestUtils
+from bifrost.src.CommonTestUtils import BaseTestUtils
 
 
 class TestLocation(BaseTestUtils):
